Refactor comparativa logic
