+++ compare-craft.html
@@ -98,7 +98,6 @@
   <script src="js/feedback-modal.js"></script>
   <!-- Nuevos módulos separados -->
   <script type="module" src="js/items-core.js"></script>
-  <script type="module" src="js/compare-logic.js"></script>
   <script type="module" src="js/compare-ui.js"></script>
 
   <script>
@@ -140,6 +139,5 @@
 
 <script src="js/storageUtils.js"></script>
 <!-- Manejadores de comparativa (guardar comparativa) -->
-<script src="js/compareHandlers.js"></script>
 </body>
 </html>
