+++ item.html
@@ -131,7 +131,6 @@
 <script src="js/storageUtils.js"></script>
 <script src="js/item-ui.js"></script>
 <script type="module" src="js/items-core.js"></script>
-<script type="module" src="js/item.js"></script>
   
   <!-- Inicialización -->
   <script>
